Only left exception part.
